added features to deck class and blackjack class
